Refactor admin views and add airline management
