+++ core/urls.py
@@ -16,14 +16,14 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from gestao import views_admin
+from gestao import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),          # Django admin padrão
     path('adm/', include('gestao.urls_admin')),  # Painel admin customizado
     path('', include('painel_cliente.urls')),    # Painel do cliente comum
-    path('adm/programas/', views_admin.admin_programas, name='admin_programas'),
-    path('adm/valor-milheiro/', views_admin.admin_valor_milheiro, name='admin_valor_milheiro'),
+    path('adm/programas/', views.admin_programas, name='admin_programas'),
+    path('adm/valor-milheiro/', views.admin_valor_milheiro, name='admin_valor_milheiro'),
 ]
