+++ gestao/templates/admin_custom/companhias.html
@@ -1,0 +1,33 @@
+{% extends "admin_custom/base_admin.html" %}
+{% block header_title %}Companhias Aéreas{% endblock %}
+{% block menu_companhias %}active{% endblock %}
+{% block content %}
+<div class="flex justify-between items-center mb-4">
+  <h2 class="text-xl font-semibold">Companhias Aéreas</h2>
+  <a href="{% url 'admin_nova_companhia' %}" class="bg-violet-700 hover:bg-violet-600 text-white px-4 py-2 rounded">Nova Companhia</a>
+</div>
+<table class="min-w-full text-sm">
+  <thead class="bg-zinc-700 text-gray-200">
+    <tr>
+      <th class="p-3 text-left font-medium">Nome</th>
+      <th class="p-3 text-left font-medium">Site</th>
+      <th class="p-3 text-left font-medium">Ações</th>
+    </tr>
+  </thead>
+  <tbody>
+    {% for companhia in companhias %}
+    <tr class="border-b border-zinc-700 hover:bg-zinc-800">
+      <td class="p-3">{{ companhia.nome }}</td>
+      <td class="p-3"><a href="{{ companhia.site_url }}" class="text-blue-400" target="_blank">{{ companhia.site_url }}</a></td>
+      <td class="p-3">
+        <a href="{% url 'admin_editar_companhia' companhia.id %}" class="text-violet-400 hover:underline">Editar</a>
+      </td>
+    </tr>
+    {% empty %}
+    <tr>
+      <td colspan="3" class="p-3 text-center text-zinc-400">Nenhuma companhia cadastrada.</td>
+    </tr>
+    {% endfor %}
+  </tbody>
+</table>
+{% endblock %}
