+++ gestao/templates/admin_custom/companhias_form.html
@@ -1,0 +1,18 @@
+{% extends "admin_custom/base_admin.html" %}
+{% block header_title %}Companhias Aéreas{% endblock %}
+{% block menu_companhias %}active{% endblock %}
+{% block content %}
+<h2 class="text-xl font-semibold mb-4">{% if form.instance.pk %}Editar Companhia{% else %}Nova Companhia{% endif %}</h2>
+<form method="post" class="space-y-4">
+  {% csrf_token %}
+  <div>
+    {{ form.nome.label_tag }}
+    {{ form.nome }}
+  </div>
+  <div>
+    {{ form.site_url.label_tag }}
+    {{ form.site_url }}
+  </div>
+  <button type="submit" class="bg-violet-700 hover:bg-violet-600 text-white px-4 py-2 rounded">Salvar</button>
+</form>
+{% endblock %}
